Fix JSON Models

+++ Qualified.Client/Data/StandardDeviationBounds.cs
@@ -5,9 +5,9 @@ namespace Qualified
 	public class StandardDeviationBounds
 	{
 		[JsonProperty("upper")]
-		public int? Upper { get; set; }
+		public double? Upper { get; set; }
 
 		[JsonProperty("lower")]
-		public int? Lower { get; set; }
+		public double? Lower { get; set; }
 	}
 }
